day3 int to long to test and check huge values

+++ src/Day3/logic.kt
@@ -2,6 +2,7 @@ package Day3
 
 typealias Position = Pair<Int, Int>
 
+
 class Player {
     var position = Position(0, 0)
 
@@ -33,24 +34,24 @@ class Player {
 
 class Grid {
 
-    var grid = mutableMapOf<String, Int>()
+    var grid = mutableMapOf<String, Long>()
     var player = Player()
 
     constructor() {
         grid[genKey(Position(0, 0))] = 1
     }
 
-    fun calculateSum(pos: Position): Int =
+    fun calculateSum(pos: Position): Long =
             (pos.first - 1..pos.first + 1).flatMap { x -> (pos.second - 1..pos.second + 1).map { y -> getEle(Pair(x, y)) } }.sum()
 
-    fun getEle(pos: Position): Int = grid.getOrDefault(genKey(pos), 0)
+    fun getEle(pos: Position): Long = grid.getOrDefault(genKey(pos), 0)
     fun genKey(pos: Position): String = "${pos.first}-${pos.second}"
 
     /**
      * Calculates closest bigger value
      */
-    fun calculateUntil(desiredVal: Int): Int {
-        var sum: Int
+    fun calculateUntil(desiredVal: Long): Long {
+        var sum: Long
         do {
             var position = player.walk() // generators?
             sum = calculateSum(position)
